split warehouse_robots_shelves_simulation.py to seperate classes - fixes #3
